src/mcp-server-youtube/pyproject.toml
[project]
name = "mcp-server-youtube"
version = "0.1.0"
description = "MCP Server providing tools to interact with YouTube."
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
readme = "README.md"
requires-python = ">=3.12"


dependencies = [
    "mcp>=1.6.0,<2.0.0",
    "starlette>=0.37.0",
    "pydantic-settings>=2.0",  
    "google-api-python-client",
    "youtube-transcript-api",
    "python-dotenv",
    "uvicorn[standard]",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx", 
    "ruff",
    "black",
    "isort",   
]
debug = [
    "debugpy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_youtube"]